server/requirements.txt
amqp==2.2.2
aniso8601==3.0.0
appnope==0.1.0
asgiref==2.3.0
asn1crypto==0.24.0
astroid==1.6.3
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.4.1
Automat==0.6.0
awscli==1.15.14
awsebcli==3.12.4
backcall==0.1.0
billiard==3.5.0.3
bintrees==2.0.7
blessed==1.14.2
boto==2.48.0
boto3==1.7.14
botocore==1.10.14
celery==4.1.0
cement==2.10.12
certifi==2018.4.16
cffi==1.11.5
channels==2.1.1
channels-api==0.4.1
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.2.2
cssselect==1.0.3
daphne==2.1.1
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.5
django-braces==1.13.0
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-filter==1.1.0
django-oauth-toolkit==1.1.0
django-querycount==0.7.0
django-redis==4.9.0
django-rest-framework-social-oauth2==1.1.0
django-ses==0.8.5
django-sslserver==0.20
django-storages==1.6.6
djangorestframework==3.8.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
drfdocs==0.0.11
first==2.0.1
flake8==3.5.0
Flask==1.0.2
future==0.16.0
gdax==1.0.6
graphene==2.1
graphql-core==2.0
graphql-relay==0.4.5
gunicorn==19.8.1
httpie==0.9.9
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython==6.3.1
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
lazy-object-proxy==1.3.1
livereload==2.5.2
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
mkdocs==0.17.3
oauthlib==2.0.7
olefile==0.45.1
packaging==17.1
parsel==1.4.0
parso==0.2.0
pathspec==0.5.6
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pip-review==1.0
pipreqs==0.4.9
popcornnotify==0.1.4
promise==2.1
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pycparser==2.18
PyDispatcher==2.0.5
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.8.4
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-decouple==3.1
python-scrapyd-api==2.1.2
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.12
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
requirements==0.1
rsa==3.4.2
Rx==1.6.1
s3transfer==0.1.13
Scrapy==1.5.0
scrapyd==1.2.0
semantic-version==2.6.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
statsd==3.2.2
tabulate==0.8.2
termcolor==1.1.0
tornado==5.0.2
traitlets==4.3.2
Twisted==18.4.0
txaio==2.10.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
virtualenv==15.2.0
w3lib==1.19.0
wcwidth==0.1.7
websocket-client==0.47.0
Werkzeug==0.14.1
whitenoise==3.3.1
wrapt==1.10.11
yarg==0.1.9
zope.interface==4.5.0
